server/requirements.txt
Flask
Flask-Cors
Flask-PyMongo
PyJWT